Changes to the components after review
changes:

app/hire/dashboard/page.tsx
app/hire/listings/page.tsx
components/common/StatusDropdown.tsx
components/features/hire/dashboard/ApplicationRow.tsx
components/features/hire/dashboard/ApplicationsTable.tsx
hooks/hire/listings/use-listings-business-logic.ts
